pyproject.toml
[project]
name = "django-qp"
version = "0.2.0"
description = "Django query parameter validation with pluggable backends (pydantic, msgspec)"
readme = "README.md"
authors = [
    { name = "Diego Margoni", email = "diegof.university@gmail.com" }
]
license = { text = "MIT" }
keywords = ["django", "drf", "pydantic", "msgspec", "query", "parameters", "validation", "fastapi"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

requires-python = ">=3.10,<3.15"
dependencies = [
    "django>=4.2,<7.0",
    "msgspec>=0.19.0",
]

[project.optional-dependencies]
pydantic = ["pydantic>=2.0"]
drf = ["djangorestframework>=3.15.2"]
all = ["pydantic>=2.0", "djangorestframework>=3.15.2"]

[dependency-groups]
dev = [
    "djangorestframework>=3.15.2",
    "pydantic>=2.0",
    "msgspec>=0.19.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.9.10",
    "ipdb>=0.13.13",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "tox>=4.0",
    "tox-uv>=1.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/django_qp"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = "test_*.py"
testpaths = ["tests"]
pythonpath = ["."]
django_find_project = false
filterwarnings = [
    "ignore:Converter .* is already registered:DeprecationWarning",
]


[tool.ruff]
line-length = 110
target-version = "py310"
lint.select = [
    "E",    # pycodestyle — style errors (indentation, spaces, etc.)
    "W",    # pycodestyle — warnings (whitespace, blank lines)
    "F",    # pyflakes — unused variables, unused imports, etc.
    "I",    # isort — import ordering
    "B",    # flake8-bugbear — common bugs and anti-patterns
    "C4",   # flake8-comprehensions — prefer list/dict/set comprehensions
    "UP",   # pyupgrade — modernise Python syntax (f-strings, union types, etc.)
    "S",    # flake8-bandit — security anti-patterns
    "T20",  # flake8-print — no print() in production code
    "SIM",  # flake8-simplify — simplify if/else, context managers, booleans
    "RUF",  # ruff-specific — extra ruff rules (redundant noqa, etc.)
    "ANN",  # flake8-annotations — mandatory type annotations on all public functions
    "D",    # pydocstyle — mandatory docstrings (Google convention, see below)
    "FBT",  # flake8-boolean-trap — no positional booleans → keyword-only arguments
]
lint.ignore = [
    "E501",  # line too long — delegated to ruff format
    "S101",  # use of assert — allowed in tests
    "UP047", # X | Y annotation syntax — not supported by the target Python version
]
lint.fixable = ["ALL"]
lint.unfixable = []
exclude = [
    "docs",
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN401", "S105"]

[tool.ruff.lint.isort]
known-first-party = ["django_qp"]

[tool.ruff.lint.pydocstyle]
# Google docstring style: Args:, Returns:, Raises:, etc.
convention = "google"

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false


[project.urls]
Homepage = "https://github.com/dfm88/django_qp"
Repository = "https://github.com/dfm88/django_qp"
Issues = "https://github.com/dfm88/django_qp/issues"
